+++ src/modules/user/infrastructure/persistence/user.repository.ts
@@ -23,7 +23,6 @@ export const UserRepository = {
   async deleteUser(id: string) {
     return await User.findByIdAndDelete(id);
   },
-
   async getUserByEmail(email: string) {
     try {
       return await User.findOne({ email });
